Fix NoClassDefFoundError on versions < 1.12 with Discord execute command
